sc5xx: move common mach- files to standard location
Move the common code associated with the sc5xx machine/architecture to
U-Boot's standard location for such files:
	arch/<ARCH>/mach-<SoC>
which, in this case, translates to:
	arch/arm/mach-sc5xx
and was originally placed at:
	board/adi/common

Additionally, in order to support this transition, modify the Kconfig
selection process to allow the user to choose the architecture
(CONFIG_ARCH_SC5XX), then choose their board (CONFIG_TARGET_SC5xx_...).
The user's choice of board automatically selects the base SoC.
