+++ arch/arm/include/asm/arch-adi/sc5xx/sc59x-shared.h
