+++ arch/arm/include/asm/arch-adi/sc5xx/soc.h
