+++ arch/arm/include/asm/arch-adi/sc5xx/spl.h
