+++ arch/arm/mach-sc5xx/Kconfig
@@ -11,8 +11,38 @@
 # All 32-bit platforms require SYS_ARM_CACHE_WRITETHROUGH
 # But it is ignored if selected here, so it must be in the defconfig
 
+if ARCH_SC5XX
+
+choice
+	prompt "SC5xx board select"
+
+config TARGET_SC598_SOM_EZKIT
+	bool "Analog Devices SC598 SOM EZ Kit"
+	select SC59X_64
+
+config TARGET_SC594_SOM_EZKIT
+	bool "Analog Devices SC594 SOM EZ Kit"
+	select SC59X
+
+config TARGET_SC589_EZKIT
+	bool "Analog Devices SC589 EZ Kit"
+	select SC58X
+
+config TARGET_SC589_MINI
+	bool "Analog Devices SC589 Mini Kit"
+	select SC58X
+
+config TARGET_SC584_EZKIT
+	bool "Analog Devices SC584 EZ Kit"
+	select SC58X
+
+config TARGET_SC573_EZKIT
+	bool "Analog Devices SC573 EZ Kit"
+	select SC57X
+endchoice
+
 config SC57X
-	bool "sc57x SoC"
+	bool
 	select SUPPORT_SPL
 	select CPU_V7A
 	select PANIC_HANG
@@ -21,7 +51,7 @@ config SC57X
 	select ADI_SC5XX_TIMER
 
 config SC58X
-	bool "sc58x SoC"
+	bool
 	select SUPPORT_SPL
 	select CPU_V7A
 	select PANIC_HANG
@@ -30,7 +60,7 @@ config SC58X
 	select ADI_SC5XX_TIMER
 
 config SC59X
-	bool "sc59x SoC"
+	bool
 	select SUPPORT_SPL
 	select CPU_V7A
 	select PANIC_HANG
@@ -39,7 +69,7 @@ config SC59X
 	select ADI_SC5XX_TIMER
 
 config SC59X_64
-	bool "sc59x SoC"
+	bool
 	select SUPPORT_SPL
 	select PANIC_HANG
 	select ARM64
@@ -312,3 +342,5 @@ source "board/adi/sc589-ezkit/Kconfig"
 source "board/adi/sc589-mini/Kconfig"
 source "board/adi/sc584-ezkit/Kconfig"
 source "board/adi/sc573-ezkit/Kconfig"
+
+endif
