+++ arch/arm/mach-sc5xx/Makefile
