+++ arch/arm/mach-sc5xx/arm32/lowlevel_init.S
