+++ arch/arm/mach-sc5xx/config.mk
