+++ arch/arm/mach-sc5xx/init/Makefile
