+++ arch/arm/mach-sc5xx/init/cguinit.h
