+++ arch/arm/mach-sc5xx/init/dmc/dmcinit.c
