+++ arch/arm/mach-sc5xx/init/dmc/mt41k512m16ha.h
