+++ arch/arm/mach-sc5xx/init/dmc/mt47h128m16rt.h
