+++ arch/arm/mach-sc5xx/init/init.h
