+++ arch/arm/mach-sc5xx/spl.c
@@ -10,7 +10,7 @@
 
 #include <asm-generic/gpio.h>
 #include <spl.h>
-#include "spl.h"
+#include <asm/arch-adi/sc5xx/spl.h>
 #include "init/init.h"
 
 static bool adi_start_uboot_proper = 1;
