--- board/adi/sc589-ezkit/config.mk
@@ -1,1 +1,0 @@
-sinclude $(srctree)/board/adi/common/init/config.mk
