+++ configs/sc573-ezkit_defconfig
@@ -1,6 +1,6 @@
 CONFIG_ARM=y
 CONFIG_SYS_ARM_CACHE_WRITETHROUGH=y
-CONFIG_TARGET_SC573_EZKIT=y
+CONFIG_ARCH_SC5XX=y
 CONFIG_SYS_TEXT_BASE=0x82200000
 CONFIG_SPL_GPIO_SUPPORT=y
 CONFIG_SPL_LIBCOMMON_SUPPORT=y
@@ -14,6 +14,7 @@ CONFIG_DM_GPIO=y
 CONFIG_SPL_TEXT_BASE=0x20080000
 CONFIG_SPL_SERIAL_SUPPORT=y
 CONFIG_SPL=y
+CONFIG_TARGET_SC573_EZKIT=y
 CONFIG_CGU0_DF_DIV=0
 CONFIG_CGU0_VCO_MULT=18
 CONFIG_CGU0_CCLK_DIV=1
