+++ configs/sc594-som-ezkit-spl_defconfig
@@ -1,6 +1,6 @@
 CONFIG_ARM=y
 CONFIG_SYS_ARM_CACHE_WRITETHROUGH=y
-CONFIG_TARGET_SC594_SOM_EZKIT=y
+CONFIG_ARCH_SC5XX=y
 CONFIG_SYS_TEXT_BASE=0xB2200000
 CONFIG_SPL_GPIO_SUPPORT=y
 CONFIG_SPL_LIBCOMMON_SUPPORT=y
@@ -15,6 +15,7 @@ CONFIG_SPL_DM_SPI=y
 CONFIG_SPL_TEXT_BASE=0x20080000
 CONFIG_SPL_SERIAL_SUPPORT=y
 CONFIG_SPL=y
+CONFIG_TARGET_SC594_SOM_EZKIT=y
 CONFIG_SPL_SPI_FLASH_SUPPORT=y
 CONFIG_SPL_SPI_SUPPORT=y
 CONFIG_SC_BOOT_MODE=2
