+++ configs/sc598-som-ezkit-spl_defconfig
@@ -1,5 +1,5 @@
 CONFIG_ARM=y
-CONFIG_TARGET_SC598_SOM_EZKIT=y
+CONFIG_ARCH_SC5XX=y
 CONFIG_SYS_TEXT_BASE=0x96000000
 CONFIG_SPL_GPIO_SUPPORT=y
 CONFIG_SPL_LIBCOMMON_SUPPORT=y
@@ -51,6 +51,7 @@ CONFIG_SPL_TEXT_BASE=0x20080000
 CONFIG_SPL_MMC_SUPPORT=y
 CONFIG_SPL_SERIAL_SUPPORT=y
 CONFIG_SPL=y
+CONFIG_TARGET_SC598_SOM_EZKIT=y
 CONFIG_SPL_SPI_FLASH_SUPPORT=y
 CONFIG_SPL_SPI_SUPPORT=y
 CONFIG_DEFAULT_DEVICE_TREE="sc598-som-ezkit"
